requirements.txt
# Core dependencies
pandas==2.1.4
numpy==1.24.3
requests==2.31.0
python-telegram-bot==20.7
schedule==1.2.0
python-dotenv==1.0.0

# Financial and technical analysis
yfinance==0.2.28
ta==0.10.2
pandas-ta==0.3.14b

# Machine Learning
xgboost==2.0.3
scikit-learn==1.3.2

# Interactive Brokers API
ibapi==10.38.0

# Data handling and utilities
pytz==2023.3
datetime
logging
json
asyncio
threading
time 